Better rendering api - fix two pass rendering strategy (both server-client and client-client)
 - Now allows either string dom id or a JSON style object of rendering
options including mountAtId
- Also - the second pass of rendering in two pass mode is intentionally
delayed by 10 ms. Seeing if that effects garbage collection.
            The two pass rendering strategy has so far shown to cause
fewer garbage collection events that block the experience of displaying
a page to the user (in chrome debugger) - likely only because it uses
less memory for the first optimized path and doesn't attach event
handlers to the top level.