Fix touch events and perf optimizations on string concatenation
I did some experiments with touch events and recorded the differences
in bubbling behavior in two iOS versions. This information could not be
found anywhere so I had to try every combination and record the results.
Diff also includes some perf optimizations on string concatenation
based engine. Got rid of odd restriction of requiring two levels deep
in the tree to mount a widget at - that's only helpful for clicks
bubbling in iOS, which are not really interesting in the first place
when you need to roll your own anyways to get rid of the 300ms delay in
click events.