Implemented nice control over memory reference freeing.
This should allow customized control over when references are deleted
to objects no longer projected onto the dom. This is nice because you
might have real time interactions and animations that can't afford to
perform this memory cleanup in the middle of an animation. Break out of
requestAnimFrame and call F.releaseMemoryReferences when you're ready
to.