Google Closure Advanced minification support! And other performance enhancements.
Jjust set minifyClosureAdvanced: true in the ProjectConfig, for
advanced minification mode - it will take a long time. Classnames will
be minified with the javascript classSets so css is tiny as well. Now
we compute a monolithicStyle.css just like we compute a
monolithicBuild.js.
Also, I reformed a tonal reorg of the file structures to break things
up into logical modules and enhanced the fax build compiler to remove
styleExports from js files since they are redundant with the
monolithicStyle.css. The example app shrinks down to about 10kb
including the entire framework and application code.