Fix issue with array children changing types.
If childList[0] originally had an instance of type X, and then it
changed to an instance of type Y, it would try to control what existed
there already - this diff makes childList more powerful in that the
children may change types and the reconciliation will be intelligent
enough to know to remove the old dom structure, free event handlers and
allocate a new structure. Also, we error if anyone uses a component
twice for rendering.